Move timezone_names file to utils folder
